+++ src/mwin/winuser.c
@@ -463,7 +463,7 @@ CreateWindowEx(DWORD dwExStyle, LPCSTR lpClassName, LPCSTR lpWindowName,
 	wp->cursor = pwp->cursor;
 	wp->cursor->usecount++;
 	wp->unmapcount = pwp->unmapcount + 1;
-	wp->id = (int)hMenu;
+	wp->id = (intptr_t)hMenu;
 	wp->gotPaintMsg = PAINT_PAINTED;
 
 	titLen = 0;
@@ -1470,11 +1470,11 @@ SetWindowPos(HWND hwnd, HWND hwndInsertAfter, int x, int y, int cx, int cy,
 	hidden = hwnd->unmapcount || (fuFlags & SWP_NOREDRAW);
 
 	if(bZorder) {
-		switch((int)hwndInsertAfter) {
-		case (int)HWND_TOP:
+		switch((intptr_t)hwndInsertAfter) {
+		case (intptr_t)HWND_TOP:
 			MwRaiseWindow(hwnd);
 			break;
-		case (int)HWND_BOTTOM:
+		case (intptr_t)HWND_BOTTOM:
 			MwLowerWindow(hwnd);
 			break;
 		default:
